Added v5.10 support
The paths are different in the legacy terminals, so the beginning
of the path to the helper dll and runtime location has to change.
I added 2 variables for this, which we'll change from the main.py
if a v5 or earlier is detected.  The function names in terminal were
changed as well as the import, to allow the path to be modified.