Bugfix/v5 fix (#10)
* Fix for working with v5.1 and lower terminals

The way paths was imported prevented helper from seeing the updated
path after it is discoverd the terminal is v5.1 and lower.

* Cleaned up imports

* Fix bug connecting to v6 after connecting to v5

Paths were only updated when connecting to a v5 terminal.  This meant that if you
first connected to a v5 terminal, then connected to a v6+ terminal, it would try
to use the v5 paths because they were never updated.